[Starts #167700724] [Chore] Restructure folder
