+++ package.json
@@ -4,15 +4,15 @@
   "description": "WayFarer is a public bus transportation booking service. (API)",
   "main": "index.js",
   "scripts": {
-    "dev": "babel-watch ./server/api/server.js",
-    "production": "babel-watch ./server/api/server.js",
-    "test": "mocha --require @babel/register ./test/*.test.js --exit --timeout 10000",
+    "dev": "babel-watch ./server/server.js",
+    "production": "babel-watch ./server/server.js",
+    "test": "mocha --require @babel/register ./server/test/*.test.js --exit --timeout 10000",
     "build": "rm -rf ./build && babel ./server -d build",
     "generate-lcov": "nyc report --reporter=text-lcov > lcov.info",
     "coveralls-coverage": "coveralls < lcov.info",
     "codeclimate-coverage": "codeclimate-test-reporter < lcov.info",
     "coverage": "nyc npm test && npm run generate-lcov && npm run coveralls-coverage && npm run codeclimate-coverage",
-    "start": "npm run build && node build/api/server.js"
+    "start": "npm run build && node build/server.js"
   },
   "repository": {
     "type": "git",
