+++ server/v1/controllers/users.js
@@ -9,6 +9,9 @@ const User = {
   * @returns {array} users objects
   */
   getAllUsers(req, res) {
+    if (!req.user.is_admin) {
+      return res.status(400).json({ status: 'error', error: 'Only admin can access this route!' });
+    }
     const users = UserModel.getAllUsers();
     if (!users.length) {
       return res.status(404).json({ status: 'Oops! No Users found!', data: [] });
@@ -22,6 +25,9 @@ const User = {
   * @returns {object} user object
   */
   getOneUser(req, res) {
+    if (!req.user.is_admin) {
+      return res.status(400).json({ status: 'error', error: 'Only admin can access this route!' });
+    }
     const userId = parseInt(req.params.userId, 10);
     const user = UserModel.getOneUser(userId);
     if (user) {
