+++ server/v1/helpers/signToken.js
